// packages/react/src/lib/components/text-passage/text-passage.d.ts
import * as React from 'react';
export declare const PdsTextPassage: React.FunctionComponent<Partial<{
    size: "default" | "sm" | "lg";
    lineLength: "default" | "sm" | "none";
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;